pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a190rithm"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "A library for machine learning and data science"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "kaggle>=1.5.13",
    "pandas>=1.3.0",
    "pyarrow>=8.0.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    "structlog>=22.1.0",
    "keyring>=23.9.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
dev = [
    "black>=22.1.0",
    "isort>=5.10.1",
    "mypy>=0.910",
    "flake8>=4.0.1",
]

[project.scripts]
kaggle-parquet = "a190rithm.tools.kaggle_downloader.cli:main"

[tool.yapf]
based_on_style = "pep8"
indent_width = 4

[tool.pytest.ini_options]
pythonpath = [
  "src"     # 兼容 scr-layout
]
addopts = "--capture=no"
norecursedirs = [
  "tests/torch",
  "a190rithm/torch"
]

[tool.pylint.main]
init-hook = "import sys; sys.path.append('src')" # 兼容 scr-layout

[tool.pyright]
include = ["a190rithm", "tests", "stubs"]
exclude = [
  "**/__pycache__",
  "a190rithm/torch",
  "tests/torch"
]

ignore = [
]

defineConstant = { DEBUG = true }
stubPath = "stubs"

reportMissingImports = true
reportMissingTypeStubs = false
reportIncompleteStub = true
reportInvalidStubStatement = true

reportGeneralTypeIssues = false

# self 都识别不了
reportUndefinedVariable = false

executionEnvironments = [
  { root = "tests", extraPaths = ["src"] },
  { root = "src" }
]
